Fix GetPage logic
Note that the outer GetPage to get the section was superflous for pages in sub dir even in Hugo 0.64, as it would fall back to a "name only" lookup.

This is still the behaviour for rel/relRef shortcode, but as this is extremely confusing for people doing a `.GetPage "somepage"` and then get a page from a totally different section instead of the expected nil (not found).